add ftruncate & usb devfile

+++ include/KF.h
@@ -8,6 +8,7 @@
 #include <unordered_map>
 #include <stdio.h>
 #include <Eigen/Dense>
+#include <glog/logging.h>
 #include "Feeds.h"
 #include "DataFusion.h"
 #include "SystemStatus.h"
@@ -46,7 +47,7 @@ public:
 
     /* 由原始数据获取位置速度信息 */
     /* res参数的first为 pos， second 为 speed */
-    void PosGetFunc(Feeds* pfeed, int Length, std::pair<int,int>& res, int& confident_flag);   
+    void PosGetFunc(Feeds* pfeed, int Length, std::pair<uint64_t,int>& res, int& confident_flag);   
 
 private:
     /* Runfunc中需要用到的迭代变量 */
